feature: 리포지토리 구현 S10P12B208-308

+++ backend/Exhale/src/main/java/com/ssafy/exhale/repository/rehabilitationRepository/CourseRepository.java
@@ -3,6 +3,5 @@ package com.ssafy.exhale.repository.rehabilitationRepository;
 import com.ssafy.exhale.domain.rehabilitation.Course;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CourseRepository extends JpaRepository<Course, Integer> {
-
+public interface CourseRepository extends JpaRepository<Course, Long> {
 }
